pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vulture"
version = "0.0.1"
authors = [
  { name="Ronan Docherty", email="ronan.docherty18@imperial.ac.uk"},
]
description = "ViT feature upsampling via CNNs"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
  "torch>=2.0.0",
  "torchvision",
  "torchmetrics>=0.10.3",
  "timm>=1.0.9",
  "numpy>=1.26.4",
  "scikit-learn>=1.6.1",
  "pillow>=11.2.0",
  "scikit-image>=0.25.2",
  "tifffile>=2025.3.30",
  "matplotlib>=3.10.0",
  "ninja>=1.11.1.1",
  "packaging",
  "interactive-seg-backend@git+https://github.com/tldr-group/interactive-seg-backend"
  ]

[project.optional-dependencies]
dev = [
  "omegaconf>=2.3.0",
  "segmentation_models_pytorch",
  "albumentations",
  "featup@git+https://github.com/mhamilton723/FeatUp",
  "pretrained_microscopy_models@git+https://github.com/nasa/pretrained-microscopy-models"
]

[project.urls]
"Homepage" = "https://github.com/tldr-group/vulture"
"Bug Tracker" = "https://github.com/tldr-group/vulture/issues"


[tool.setuptools]
packages = ["vulture", "vulture.models"]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]